requirements.txt
websocket-client>=0.44.0
requests>=2.13.0
